Move files into appropriate crates
