+++ dams-key-server/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "dams-key-server"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+dams = { version = "0.1.0", path = "../dams" }
+futures = "0"
+serde = "1"
+serde_with = "1"
+structopt = "0"
+thiserror = "1"
+tracing = "0"
+tracing-futures = "0"
+tracing-subscriber = { version = "0", features = ["env-filter"] }
