+++ dams-key-server/src/main.rs
@@ -2,13 +2,13 @@ use structopt::StructOpt;
 use tracing::error;
 use tracing_subscriber::EnvFilter;
 
-pub(crate) mod client;
+pub mod policy_engine;
 
+mod cli;
 mod server;
 
 #[derive(Debug, StructOpt)]
 pub enum Cli {
-    Client(da_mgmt::client::Cli),
     Server(da_mgmt::server::Cli),
 }
 
@@ -17,10 +17,9 @@ pub async fn main() {
     let filter = EnvFilter::try_new("info,sqlx::query=warn").unwrap();
     tracing_subscriber::fmt().with_env_filter(filter).init();
 
-    use Cli::{Client, Server};
+    use Cli::Server;
     let result = match Cli::from_args() {
         Server(cli) => server::main_with_cli(cli).await,
-        Client(cli) => client::main_with_cli(cli).await,
     };
     if let Err(e) = result {
         error!("{}, caused by: {}", e, e.root_cause());
