+++ dams-remote-client/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "dams-remote-client"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+dams = { version = "0.1.0", path = "../dams" }
+thiserror = "1"
