+++ dams-remote-client/src/lib.rs
@@ -1,0 +1,12 @@
+//! This crate is an implementation of a remote client to a key management system.
+//!
+//!
+#![warn(missing_debug_implementations)]
+#![warn(unused_results)]
+#![warn(future_incompatible)]
+#![warn(unused)]
+#![forbid(rustdoc::broken_intra_doc_links)]
+
+use std::fmt;
+
+pub mod api;
