+++ dams/src/key_mgmt/client.rs
@@ -1,4 +1,4 @@
-use crate::{client::Config, protocol, transport::KeyMgmtAddress};
+use crate::{config::client::Config, protocol, transport::KeyMgmtAddress};
 #[cfg(feature = "allow_explicit_certificate_trust")]
 use anyhow::Context;
 use async_trait::async_trait;
