+++ dams/src/lib.rs
@@ -1,4 +1,4 @@
-//! This crate is an implementation of a key management system
+//! This crate is an implementation of a local client to a key management system.
 #![warn(missing_debug_implementations)]
 #![warn(unused_results)]
 #![warn(future_incompatible)]
@@ -7,8 +7,16 @@
 
 use std::fmt;
 
-pub mod client;
-pub mod server;
+pub mod blockchain;
+pub mod config;
+pub mod defaults;
+pub mod key_mgmt;
+pub mod keys;
+pub mod protocol;
+pub mod timeout;
+pub mod transaction;
+pub mod transport;
+
 
 /// Logs used to verify that an operation completed in the integration tests.
 #[derive(Debug)]
